+++ lib/game.rb
@@ -1,0 +1,9 @@
+class Game
+
+    attr_accessor :state
+
+    def initialize(state=['', '', '', '', '', '', '', '', ''])
+      @state = state
+    end
+    
+end
