+++ lib/turn.rb
@@ -1,0 +1,13 @@
+class Turn
+
+    attr_accessor :turn
+
+    def initialize
+      @turn = "X"
+    end
+
+    def switch_turn
+     @turn = @turn == "X" ? "O" : "X"
+    end
+
+end
